requirements.txt
pyTelegramBotAPI==4.14.0
python-telegram-bot==20.7
pytube==15.0.0
telebot==0.0.5
fastapi==0.115.8